+++ Api/.flaskenv
@@ -1,4 +1,4 @@
 DEBUG = True
-FLASK_ENV=development
-FLASK_APP=main.py
+FLASK_ENV = development
+FLASK_APP = main.py
