+++ client/src/assets/css/global.css
@@ -1,0 +1,18 @@
+/* Provide sufficient contrast against white background */
+.btn:focus, .btn:active:focus, .btn-link.nav-link:focus, .form-control:focus, .form-check-input:focus {
+    box-shadow: 0 0 0 0.1rem gray, 0 0 0 0.25rem #258cfb;
+  }
+  
+  code {
+    color: #000000;
+  }
+  
+  .btn-primary {
+    color: #2f9459;
+    background-color: #1b6fc242;
+    border-color: #18aca0;
+  }
+  
+  html, body {
+    height:100%;
+  }
