--- client/src/components/pages/functions/MousePostion.css
@@ -1,38 +1,0 @@
-.container {
-    margin: 0;
-    display: inline;
-    width: auto;
-    
-}
-
-.cursor {
-    position: absolute;
-    
-    border-radius: 50%;
-    transform: "translate(-50%, -50%)";
-    pointer-events:none;
-}
-
-.cursors .cursor:nth-child(1) {
-    /*background-color: #3a26fd;*/
-    background-image: repeating-radial-gradient(circle at 17% 32%, rgba(240, 29, 29), rgb(236, 107, 107) 0.00085px);
-    animation: tv-static 25s linear infinite;
-    height: 9px;
-    width: 9px;
-    border-radius: 50%;
-    z-index: 99999;
-    opacity: 1;
-    filter: contrast(1.75);
-    filter: brightness(200%);
-}
-.cursors .cursor:nth-child(2) {
-    background-image: repeating-radial-gradient(circle at 17% 32%, rgb(240, 29, 29), rgb(236, 107, 107) 0.0085px);
-    animation: tv-static 25s linear infinite;
-    z-index: 99998;
-    height: 15px;
-    width: 15px;
-    border-radius: 50%;
-    opacity: 0.1;
-    filter: contrast(1.75);
-    filter: brightness(200%);
-}
