+++ client/src/hooks/MousePostion.js
@@ -1,5 +1,5 @@
 import React,{ Component, createRef } from "react";
-import './MousePostion.css'
+
 
 export class MousePosition extends Component{
