+++ client/src/layouts/footer/Footer.js
@@ -1,4 +1,4 @@
-import React,{ Component } from "react";
+import React from "react";
 import { MDBFooter } from 'mdb-react-ui-kit';
 import './Footer.css';
