+++ client/src/layouts/navbar/NavMenu.css
@@ -17,8 +17,6 @@ html {
 .logo_img {
   
   position: absolute;
-  left: 1%;
-  top: 0;
 }
 
 img:first-of-type{
