+++ client/src/pages/animation.js
@@ -1,0 +1,8 @@
+export default function animation(){
+
+    return(
+        <div className="container">
+
+        </div>
+    )
+}
